Fix ebpf verifier load + add exponential backoff
The verifier because of the nested ifs was having a hard time retaining
the bounds of the registers and assessing the sane access to memory.
Indeed, when a register is rewritten after a "if", the knowledge brought
by the condition is lost in the verifier. To fix that, this commit adds
functions that allow to flatten the nested ifs and prevent the compiler
to erase register values right after testing them.